+++ just_audio/lib/just_audio.dart
@@ -37,6 +37,9 @@ final _uuid = Uuid();
 /// You must call [dispose] to release the resources used by this player,
 /// including any temporary files created to cache assets.
 class AudioPlayer {
+  /// The user agent to set on all HTTP requests.
+  final String _userAgent;
+
   /// This is `true` when the audio player needs to engage the native platform
   /// side of the plugin to decode or play audio, and is `false` when the native
   /// resources are not needed (i.e. after initial instantiation and after [stop]).
@@ -100,14 +103,38 @@ class AudioPlayer {
   bool _automaticallyWaitsToMinimizeStalling = true;
   bool _playInterrupted = false;
   AndroidAudioAttributes _androidAudioAttributes;
+  bool _androidApplyAudioAttributes;
+  bool _handleAudioSessionActivation;
 
-  /// Creates an [AudioPlayer]. The player will automatically pause/duck and
-  /// resume/unduck when audio interruptions occur (e.g. a phone call) or when
-  /// headphones are unplugged. If you wish to handle audio interruptions
-  /// manually, set [handleInterruptions] to `false` and interface directly
-  /// with the audio session via the
-  /// [audio_session](https://pub.dev/packages/audio_session) package.
-  AudioPlayer({bool handleInterruptions = true}) : _id = _uuid.v4() {
+  /// Creates an [AudioPlayer].
+  ///
+  /// If [userAgent] is specified, it will be included in the header of all HTTP
+  /// requests on Android, iOS and macOS to identify your agent to the server.
+  /// If set, just_audio will create a cleartext local HTTP proxy on your device
+  /// to forward HTTP requests with headers included. If [userAgent] is not
+  /// specified, this will default to Apple's Core Audio user agent on iOS/macOS
+  /// and to just_audio's own user agent on Android. On Web, the browser will
+  /// override any specified user-agent string with its own.
+  ///
+  /// The player will automatically pause/duck and resume/unduck when audio
+  /// interruptions occur (e.g. a phone call) or when headphones are unplugged.
+  /// If you wish to handle audio interruptions manually, set
+  /// [handleInterruptions] to `false` and interface directly with the audio
+  /// session via the [audio_session](https://pub.dev/packages/audio_session)
+  /// package. If you do not wish just_audio to automatically activate the audio
+  /// session when playing audio, set [handleAudioSessionActivation] to `false`.
+  /// If you do not want just_audio to respect the global
+  /// [AndroidAudioAttributes] configured by audio_session, set
+  /// [androidApplyAudioAttributes] to `false`.
+  AudioPlayer({
+    String userAgent,
+    bool handleInterruptions = true,
+    bool androidApplyAudioAttributes = true,
+    bool handleAudioSessionActivation = true,
+  })  : _id = _uuid.v4(),
+        _userAgent = userAgent,
+        _androidApplyAudioAttributes = androidApplyAudioAttributes,
+        _handleAudioSessionActivation = handleAudioSessionActivation {
     _idlePlatform = _IdleAudioPlayer(id: _id, sequenceStream: sequenceStream);
     _playbackEventSubject.add(_playbackEvent = PlaybackEvent());
     _processingStateSubject.addStream(playbackEventStream
@@ -140,7 +167,7 @@ class AudioPlayer {
       (sequence, shuffleIndices, currentIndex, shuffleModeEnabled, loopMode) {
         if (sequence == null) return null;
         if (currentIndex == null) currentIndex = 0;
-        currentIndex = min(sequence.length - 1, max(0, currentIndex));
+        currentIndex = max(min(sequence.length - 1, max(0, currentIndex)), 0);
         return SequenceState(
           sequence,
           currentIndex,
@@ -162,13 +189,15 @@ class AudioPlayer {
     _setPlatformActive(false);
     _sequenceSubject.add(null);
     // Respond to changes to AndroidAudioAttributes configuration.
-    AudioSession.instance.then((audioSession) {
-      audioSession.configurationStream
-          .map((conf) => conf?.androidAudioAttributes)
-          .where((attributes) => attributes != null)
-          .distinct()
-          .listen(setAndroidAudioAttributes);
-    });
+    if (androidApplyAudioAttributes) {
+      AudioSession.instance.then((audioSession) {
+        audioSession.configurationStream
+            .map((conf) => conf?.androidAudioAttributes)
+            .where((attributes) => attributes != null)
+            .distinct()
+            .listen(setAndroidAudioAttributes);
+      });
+    }
     if (handleInterruptions) {
       AudioSession.instance.then((session) {
         session.becomingNoisyEventStream.listen((_) {
@@ -497,6 +526,8 @@ class AudioPlayer {
 
   /// Convenience method to set the audio source to a URL with optional headers,
   /// preloaded by default, with an initial position of zero by default.
+  /// If headers are set, just_audio will create a cleartext local HTTP proxy on
+  /// your device to forward HTTP requests with headers included.
   ///
   /// This is equivalent to:
   ///
@@ -519,7 +550,7 @@ class AudioPlayer {
   /// default, with an initial position of zero by default.
   ///
   /// ```
-  /// setAudioSource(AudioSource.uri(Uri.file(filePath)), headers: headers,
+  /// setAudioSource(AudioSource.uri(Uri.file(filePath)),
   ///     initialPosition: Duration.zero, preload: true);
   /// ```
   ///
@@ -537,7 +568,7 @@ class AudioPlayer {
   ///
   /// ```
   /// setAudioSource(AudioSource.uri(Uri.parse('asset:///$assetPath')),
-  ///     headers: headers, initialPosition: Duration.zero, preload: true);
+  ///     initialPosition: Duration.zero, preload: true);
   /// ```
   ///
   /// See [setAudioSource] for a detailed explanation of the options.
@@ -645,7 +676,7 @@ class AudioPlayer {
   Future<Duration> _load(AudioPlayerPlatform platform, AudioSource source,
       {_InitialSeekValues initialSeekValues}) async {
     try {
-      if (!kIsWeb && source._requiresProxy) {
+      if (!kIsWeb && (source._requiresProxy || _userAgent != null)) {
         if (_proxy == null) {
           _proxy = _ProxyHttpServer();
           await _proxy.start();
@@ -732,7 +763,7 @@ class AudioPlayer {
     _playbackEventSubject.add(_playbackEvent);
     final playCompleter = Completer();
     final audioSession = await AudioSession.instance;
-    if (await audioSession.setActive(true)) {
+    if (!_handleAudioSessionActivation || await audioSession.setActive(true)) {
       // TODO: rewrite this to more cleanly handle simultaneous load/play
       // requests which each may result in platform play requests.
       final requireActive = _audioSource != null;
@@ -1054,10 +1085,12 @@ class AudioPlayer {
         final playing = this.playing;
         // To avoid a glitch in ExoPlayer, ensure that any requested audio
         // attributes are set before loading the audio source.
-        final audioSession = await AudioSession.instance;
-        if (_androidAudioAttributes == null) {
-          _androidAudioAttributes =
-              audioSession.configuration?.androidAudioAttributes;
+        if (_androidApplyAudioAttributes) {
+          final audioSession = await AudioSession.instance;
+          if (_androidAudioAttributes == null) {
+            _androidAudioAttributes =
+                audioSession.configuration?.androidAudioAttributes;
+          }
         }
         if (_androidAudioAttributes != null) {
           await _internalSetAndroidAudioAttributes(
@@ -1404,7 +1437,8 @@ class SequenceState {
       this.shuffleModeEnabled, this.loopMode);
 
   /// The current source in the sequence.
-  IndexedAudioSource get currentSource => sequence[currentIndex];
+  IndexedAudioSource get currentSource =>
+      sequence.isEmpty ? null : sequence[currentIndex];
 
   /// The effective sequence. This is equivalent to [sequence]. If
   /// [shuffleModeEnabled] is true, this is modulated by [shuffleIndices].
@@ -1428,7 +1462,13 @@ class _ProxyHttpServer {
   /// called only after [start] has completed.
   Uri addUriAudioSource(UriAudioSource source) {
     final uri = source.uri;
-    final headers = source.headers?.cast<String, String>();
+    final headers = <String, String>{};
+    if (source.headers != null) {
+      headers.addAll(source.headers.cast<String, String>());
+    }
+    if (source._player._userAgent != null) {
+      headers['user-agent'] = source._player._userAgent;
+    }
     final path = _requestKey(uri);
     _handlerMap[path] = _proxyHandlerForUri(uri, headers);
     return uri.replace(
@@ -1538,6 +1578,9 @@ abstract class AudioSource {
   /// stream type on Android. If you know in advance what type of audio stream
   /// it is, you should instantiate [DashAudioSource] or [HlsAudioSource]
   /// directly.
+  ///
+  /// If headers are set, just_audio will create a cleartext local HTTP proxy on
+  /// your device to forward HTTP requests with headers included.
   static AudioSource uri(Uri uri, {Map headers, dynamic tag}) {
     bool hasExtension(Uri uri, String extension) =>
         uri.path.toLowerCase().endsWith('.$extension') ||
@@ -1617,7 +1660,7 @@ abstract class UriAudioSource extends IndexedAudioSource {
     await super._setup(player);
     if (uri.scheme == 'asset') {
       _overrideUri = await _loadAsset(uri.path.replaceFirst(RegExp(r'^/'), ''));
-    } else if (headers != null) {
+    } else if (headers != null || player._userAgent != null) {
       _overrideUri = player._proxy.addUriAudioSource(this);
     }
   }
@@ -1683,6 +1726,9 @@ abstract class UriAudioSource extends IndexedAudioSource {
 ///
 /// On platforms except for the web, the supplied [headers] will be passed with
 /// the HTTP(S) request.
+///
+/// If headers are set, just_audio will create a cleartext local HTTP proxy on
+/// your device to forward HTTP requests with headers included.
 class ProgressiveAudioSource extends UriAudioSource {
   ProgressiveAudioSource(Uri uri, {Map headers, dynamic tag, Duration duration})
       : super(uri, headers: headers, tag: tag, duration: duration);
@@ -1703,6 +1749,9 @@ class ProgressiveAudioSource extends UriAudioSource {
 /// On platforms except for the web, the supplied [headers] will be passed with
 /// the HTTP(S) request. Currently headers are not recursively applied to items
 /// the HTTP(S) request. Currently headers are not applied recursively.
+///
+/// If headers are set, just_audio will create a cleartext local HTTP proxy on
+/// your device to forward HTTP requests with headers included.
 class DashAudioSource extends UriAudioSource {
   DashAudioSource(Uri uri, {Map headers, dynamic tag, Duration duration})
       : super(uri, headers: headers, tag: tag, duration: duration);
@@ -1722,6 +1771,9 @@ class DashAudioSource extends UriAudioSource {
 ///
 /// On platforms except for the web, the supplied [headers] will be passed with
 /// the HTTP(S) request. Currently headers are not applied recursively.
+///
+/// If headers are set, just_audio will create a cleartext local HTTP proxy on
+/// your device to forward HTTP requests with headers included.
 class HlsAudioSource extends UriAudioSource {
   HlsAudioSource(Uri uri, {Map headers, dynamic tag, Duration duration})
       : super(uri, headers: headers, tag: tag, duration: duration);
@@ -2092,6 +2144,12 @@ class LockCachingAudioSource extends StreamAudioSource {
   int _progress = 0;
   final _requests = <_StreamingByteRangeRequest>[];
 
+  /// Creates a [LockCachingAudioSource] to that provides [uri] to the player
+  /// while simultaneously caching it to [cacheFile]. If no cache file is
+  /// supplied, just_audio will allocate a cache file internally.
+  ///
+  /// If headers are set, just_audio will create a cleartext local HTTP proxy on
+  /// your device to forward HTTP requests with headers included.
   LockCachingAudioSource(
     this.uri, {
     this.headers,
@@ -2127,14 +2185,32 @@ class LockCachingAudioSource extends StreamAudioSource {
     }
   }
 
+  /// Start downloading the whole audio file to the cache and fulfill byte-range
+  /// requests during the download. There are 3 scenarios:
+  ///
+  /// 1. If the byte range request falls entirely within the cache region, it is
+  /// fulfilled from the cache.
+  /// 2. If the byte range request overlaps the cached region, the first part is
+  /// fulfilled from the cache, and the region beyond the cache is fulfilled
+  /// from a memory buffer of the downloaded data.
+  /// 3. If the byte range request is entirely outside the cached region, a
+  /// separate HTTP request is made to fulfill it while the download of the
+  /// entire file continues in parallel.
   Future<HttpClientResponse> _fetch() async {
+    final cacheFile = await _cacheFile;
+    final partialCacheFile = await _partialCacheFile;
+    final mimeType = await _readCachedMimeType();
+
+    File getEffectiveCacheFile() =>
+        partialCacheFile.existsSync() ? partialCacheFile : cacheFile;
+
     HttpClient httpClient = HttpClient();
-    final request = await httpClient.getUrl(uri);
+    final httpRequest = await httpClient.getUrl(uri);
     if (headers != null) {
-      request.headers.clear();
-      headers.forEach((name, value) => request.headers.set(name, value));
+      httpRequest.headers.clear();
+      headers.forEach((name, value) => httpRequest.headers.set(name, value));
     }
-    final response = await request.close();
+    final response = await httpRequest.close();
     if (response.statusCode != 200) {
       httpClient.close();
       throw Exception('HTTP Status Error: ${response.statusCode}');
@@ -2144,28 +2220,92 @@ class LockCachingAudioSource extends StreamAudioSource {
     // ignore: close_sinks
     final sink = (await _partialCacheFile).openWrite();
     var sourceLength = response.contentLength;
+    final inProgressResponses = <_InProgressCacheResponse>[];
     StreamSubscription subscription;
-    subscription = response.listen((data) {
+    //int percentProgress = 0;
+    subscription = response.listen((data) async {
       _progress += data.length;
+      //int newPercentProgress = 100 * _progress ~/ sourceLength;
+      //if (newPercentProgress != percentProgress) {
+      //  percentProgress = newPercentProgress;
+      //  print("### Progress: $percentProgress%");
+      //}
       sink.add(data);
-      final readyRequests = _requests
-          .where((request) => request.end ?? sourceLength <= _progress)
-          .toList();
-      if (readyRequests.isEmpty) return;
-      sink.flush().then((_) async {
-        for (var request in readyRequests) {
-          _requests.remove(request);
-          final start = request.start ?? 0;
-          final end = request.end ?? sourceLength;
+      final readyRequests =
+          _requests.where((request) => (request.start) < _progress).toList();
+      final notReadyRequests =
+          _requests.where((request) => (request.start) >= _progress).toList();
+      // Add this live data to any responses in progress.
+      for (var cacheResponse in inProgressResponses) {
+        if (_progress >= cacheResponse.end) {
+          // We've received enough data to fulfill the byte range request.
+          cacheResponse.controller.add(
+              data.sublist(0, data.length - (_progress - cacheResponse.end)));
+          cacheResponse.controller.close();
+        } else {
+          cacheResponse.controller.add(data);
+        }
+      }
+      if (_requests.isEmpty) return;
+      // Prevent further data coming from the HTTP source until we have set up
+      // an entry in inProgressResponses to continue receiving live HTTP data.
+      subscription.pause();
+      await sink.flush();
+      // Process any requests that start within the cache.
+      for (var request in readyRequests) {
+        _requests.remove(request);
+        final start = request.start ?? 0;
+        final end = request.end ?? sourceLength;
+        Stream<List<int>> responseStream;
+        if (end <= _progress) {
+          responseStream = getEffectiveCacheFile().openRead(start, end);
+        } else {
+          final cacheResponse = _InProgressCacheResponse(end: end);
+          inProgressResponses.add(cacheResponse);
+          responseStream = Rx.concatEager([
+            // NOTE: The cache file part of the stream must not overlap with
+            // the live part. "_progress" should
+            // to the cache file at the time
+            getEffectiveCacheFile().openRead(start, _progress),
+            cacheResponse.controller.stream,
+          ]);
+        }
+        request.complete(StreamAudioResponse(
+          sourceLength: sourceLength,
+          contentLength: end - start,
+          offset: start,
+          contentType: mimeType,
+          stream: responseStream,
+        ));
+      }
+      subscription.resume();
+      // Process any requests that start beyond the cache.
+      for (var request in notReadyRequests) {
+        _requests.remove(request);
+        final start = request.start ?? 0;
+        final end = request.end ?? sourceLength;
+        httpClient.getUrl(uri).then((httpRequest) async {
+          if (headers != null) {
+            httpRequest.headers.clear();
+            headers
+                .forEach((name, value) => httpRequest.headers.set(name, value));
+          }
+          httpRequest.headers
+              .set(HttpHeaders.rangeHeader, 'bytes=$start-${end - 1}');
+          final response = await httpRequest.close();
+          if (response.statusCode != 206) {
+            httpClient.close();
+            throw Exception('HTTP Status Error: ${response.statusCode}');
+          }
           request.complete(StreamAudioResponse(
             sourceLength: sourceLength,
             contentLength: end - start,
             offset: start,
-            contentType: await _readCachedMimeType(),
-            stream: (await _effectiveCacheFile).openRead(start, end),
+            contentType: mimeType,
+            stream: response,
           ));
-        }
-      });
+        });
+      }
     }, onDone: () async {
       (await _partialCacheFile).renameSync((await _cacheFile).path);
       await subscription.cancel();
@@ -2177,9 +2317,6 @@ class LockCachingAudioSource extends StreamAudioSource {
     });
     return response;
   }
-
-  Future<File> get _effectiveCacheFile async =>
-      (await _partialCacheFile).existsSync() ? _partialCacheFile : _cacheFile;
 
   @override
   Future<StreamAudioResponse> request([int start, int end]) async {
@@ -2198,11 +2335,29 @@ class LockCachingAudioSource extends StreamAudioSource {
     }
     final byteRangeRequest = _StreamingByteRangeRequest(start, end);
     _requests.add(byteRangeRequest);
-    if (_response == null) {
-      _response = _fetch();
-    } else {}
+    _response ??= _fetch();
     return byteRangeRequest.future;
   }
+}
+
+/// When a byte range request on a [LockCachingAudioSource] overlaps partially
+/// with the cache file and partially with the live HTTP stream, the consumer
+/// needs to first consume the cached part before the live part. This class
+/// provides a place to buffer the live part until the consumer reaches it, and
+/// also keeps track of the [end] of the byte range so that the producer knows
+/// when to stop adding data.
+class _InProgressCacheResponse {
+  // NOTE: This isn't necessarily memory efficient. Since the entire audio file
+  // will likely be downloaded at a faster rate than the rate at which the
+  // player is consuming audio data, it is also likely that this buffered data
+  // will never be used.
+  // TODO: Improve this code.
+  // ignore: close_sinks
+  final controller = ReplaySubject<List<int>>();
+  final int end;
+  _InProgressCacheResponse({
+    @required this.end,
+  });
 }
 
 /// Request parameters for a [StreamingAudioSource].