+++ just_audio/test/just_audio_test.dart
@@ -233,7 +233,7 @@ void runTests() {
 
   test('speed', () async {
     final player = AudioPlayer();
-    final duration = await player.setUrl('https://foo.foo/foo.mp3');
+    /*final duration =*/ await player.setUrl('https://foo.foo/foo.mp3');
     final period1 = Duration(seconds: 2);
     final period2 = Duration(seconds: 2);
     final speed1 = 0.75;
@@ -253,7 +253,7 @@ void runTests() {
 
   test('positionStream', () async {
     final player = AudioPlayer();
-    final duration = await player.setUrl('https://foo.foo/foo.mp3');
+    /*final duration =*/ await player.setUrl('https://foo.foo/foo.mp3');
     final period = Duration(seconds: 3);
     final position1 = period;
     final position2 = position1 + period;
@@ -299,7 +299,7 @@ void runTests() {
   test('icyMetadata', () async {
     final player = AudioPlayer();
     expect(player.icyMetadata, equals(null));
-    final duration = await player.setUrl('https://foo.foo/foo.mp3');
+    /*final duration =*/ await player.setUrl('https://foo.foo/foo.mp3');
     player.play();
     expect(player.icyMetadata.headers.genre, equals(icyMetadata.headers.genre));
     expect((await player.icyMetadataStream.first).headers.genre,
@@ -415,6 +415,41 @@ void runTests() {
     expect(source2.sequence.map((s) => s.tag as String), equals(['e']));
     await source2.clear();
     expect(source2.sequence.map((s) => s.tag as String), equals([]));
+    await player.dispose();
+  });
+
+  test('sequence-state', () async {
+    final player = AudioPlayer();
+    expect(player.sequenceState, equals(null));
+    final playlist = ConcatenatingAudioSource(
+      children: [
+        AudioSource.uri(
+          Uri.parse("https://bar.bar/foo.mp3"),
+          tag: 'foo',
+        ),
+        AudioSource.uri(
+          Uri.parse("https://baz.baz/bar.mp3"),
+          tag: 'bar',
+        ),
+      ],
+    );
+    await player.setAudioSource(playlist);
+    expect(player.sequenceState?.sequence, equals(playlist.children));
+    expect(player.sequenceState?.currentIndex, equals(0));
+    expect(player.sequenceState?.currentSource, equals(playlist.children[0]));
+    await player.seekToNext();
+    expect(player.sequenceState?.sequence, equals(playlist.children));
+    expect(player.sequenceState?.currentIndex, equals(1));
+    expect(player.sequenceState?.currentSource, equals(playlist.children[1]));
+    await playlist.removeAt(1);
+    expect(player.sequenceState?.sequence, equals(playlist.children));
+    expect(player.sequenceState?.currentIndex, equals(0));
+    expect(player.sequenceState?.currentSource, equals(playlist.children[0]));
+    await playlist.removeAt(0);
+    expect(player.sequenceState?.sequence, equals(playlist.children));
+    // expecting 0 here may change in a future version.
+    expect(player.sequenceState?.currentIndex, equals(0));
+    expect(player.sequenceState?.currentSource, equals(null));
     await player.dispose();
   });
 
@@ -830,13 +865,9 @@ class MockAudioPlayer implements AudioPlayerPlatform {
   Duration _updatePosition = Duration.zero;
   DateTime _updateTime = DateTime.now();
   Duration _duration = audioSourceDuration;
-  int _currentIndex;
   int _index;
   var _playing = false;
   var _speed = 1.0;
-  var _volume = 1.0;
-  var _loopMode = LoopModeMessage.off;
-  var _shuffleModeEnabled = false;
   Completer<dynamic> _playCompleter;
   Timer _playTimer;
 
@@ -921,14 +952,12 @@ class MockAudioPlayer implements AudioPlayerPlatform {
 
   @override
   Future<SetLoopModeResponse> setLoopMode(SetLoopModeRequest request) async {
-    _loopMode = request.loopMode;
     return SetLoopModeResponse();
   }
 
   @override
   Future<SetShuffleModeResponse> setShuffleMode(
       SetShuffleModeRequest request) async {
-    _shuffleModeEnabled = request.shuffleMode == ShuffleModeMessage.all;
     return SetShuffleModeResponse();
   }
 
@@ -947,7 +976,6 @@ class MockAudioPlayer implements AudioPlayerPlatform {
 
   @override
   Future<SetVolumeResponse> setVolume(SetVolumeRequest request) async {
-    _volume = request.volume;
     return SetVolumeResponse();
   }